split the code and fix the style
